requirements.txt
PySide6
pynput
mss
pytesseract
python-dotenv
openai
Pillow
